frontend/Resume_parser/Resume_parser/backend/requirements.txt
fastapi
uvicorn
python-multipart
firebase-admin
groq
pymupdf
python-docx
python-dotenv
